web-27-new-layout-missing-modals: moved shop infor call from main page to trip creator list modal
