+++ wetmap/src/components/newModals/shopModal/index.tsx
@@ -15,7 +15,8 @@ type ShopModalProps = Partial<ModalHandleProps>;
 export default function ShopModal(props: ShopModalProps) {
   const { selectedShop, setSelectedShop } = useContext(SelectedShopContext);
   const { profile } = useContext(UserProfileContext);
-  const [isMyShop, setIsMyShop] = useState<boolean>(false);  const [isPartnerAccount, setIsPartnerAccount] = useState(false);
+  const [isMyShop, setIsMyShop] = useState<boolean>(false);
+  const [isPartnerAccount, setIsPartnerAccount] = useState(false);
   const [itineraryList, setItineraryList] = useState<ItineraryItem[]>([]);
   const [selectedID, setSelectedID] = useState<number>(0);
   const { modalShow } = useContext(ModalContext);
@@ -28,8 +29,8 @@ export default function ShopModal(props: ShopModalProps) {
       setIsPartnerAccount(true);
     }
     if (
-      (profile?.partnerAccount) &&
-      (selectedShop?.userId === profile.UserID)
+      (profile?.partnerAccount)
+      && (selectedShop?.userId === profile.UserID)
     ) {
       setIsMyShop(true);
     } else {
@@ -72,7 +73,7 @@ export default function ShopModal(props: ShopModalProps) {
   const openTripCreatorList = async () => {
     modalShow(TripCreatorListModal, {
       keepPreviousModal: true,
-      size: 'medium',
+      size:              'medium',
     });
   };
